Renamed printers2 and sentence2 to printers and sentence.  Also fixed all references
