+++ proof.py
@@ -1,10 +1,10 @@
 from collections import deque
 
 import inference
-import sentence2 as sentence
+import sentence
 import line
 import util
-import printers2 as printers
+import printers
 
 class Proof:
     '''
